fix int > size_t: ft_strlen.c

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: isa-viei <isa-viei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/28 01:56:23 by isa-viei          #+#    #+#             */
-/*   Updated: 2021/06/03 02:38:52 by isa-viei         ###   ########.fr       */
+/*   Updated: 2021/06/04 01:14:17 by isa-viei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
